+++ Source/Core/Utilities/NovusUtilities.pas
@@ -23,8 +23,6 @@ Type
     class function FindStringListValue(const Strings: tstringlist;
       Name: String): String;
     class function FindFileSize(Afile: String): Integer;
-//    class function JustFilename(const PathName: String): String;
-//    class function JustPathname(const PathName: ShortString): String;
     class function GetPropertyasClass(aObject: TObject;
       aPropertyName: string): TObject;
     class function IsProperty(aObject: TObject; aPropertyName: string): Boolean;
@@ -83,46 +81,7 @@ begin
 
   SysUtils.FindClose(SR);
 end;
-(*
-class function TNovusUtilities.JustFilename(const PathName: String): String;
-var
-  I: Longint;
-begin
-  Result := '';
-  if PathName = '' then
-    Exit;
-  I := Succ(Length(PathName));
-  repeat
-    Dec(I);
-  until (I = 0) or (PathName[I] in DosDelimSet);
-  Result := Copy(PathName, Succ(I), StMaxFileLen);
-end;
 
-class function TNovusUtilities.JustPathname(const PathName
-  : ShortString): String;
-Var
-  I: Integer;
-begin
-  I := Succ(Length(PathName));
-  repeat
-    Dec(I);
-  until (I = 0) or (PathName[I] in DosDelimSet);
-
-  if I = 0 then
-    SetLength(Result, 0)
-  else if I = 1 then
-    Result := PathName[1]
-  else if (PathName[I] = '\') then
-  begin
-    if PathName[Pred(I)] = ':' then
-      Result := Copy(PathName, 1, I)
-    else
-      Result := Copy(PathName, 1, Pred(I));
-  end
-  else
-    Result := Copy(PathName, 1, I);
-end;
-*)
 
 class function TNovusUtilities.CopyObject(Src, Dest: TObject;
   Related: Boolean = FALSE): Boolean;